0.3.0 (#140)
# General notes

This release is been mostly focus on automating the release to snapcraft. 

Also, now the spellchecker works with a local version of hunspell installed (if located under /usr/share/hunspell/)

The icon has been also updated to the latest MS one.

A couple of new features added to be able to start the app minimised and having the webDebug option in the tray menu.

# Detail changes

* #92 Using gitflow from now on for releases (having master and develop branches) (Thanks to @julian-alarcon)
* #74 automating the snap release (Thanks to @julian-alarcon and @benyanke)
* #121 Fixing snap and camera not working together (Thanks to @julian-alarcon)
* #118, #113 and #119 Unpacking the spellchecker dictionaries and adding the ones installed in hunspell (Thanks to @vadimi and @noderat )
* #134 Updating the MS icon (Thanks to @julian-alarcon)
* #91 removing the apparmor=DENIED messages in syslog for snap build (Thanks to @julian-alarcon)
* #137 Adding the --minimized option to start the app minimized (Thanks to @didster)
* #135 Adding webDebug to the tray menu options (Thanks to @Fmstrat)

Also:
* #109 At the moment the propagation of 'about:blank' links has been disabled to minimise annoyance from this issue.

And the packages versions has been updated a bit, as we got a blocker for moving to version 5 or electron.


# Thanks

To everyone involved from people reporting issues, testing solutions and implementing them. You all make this possible.