Add video menu entry that allows control of video elements
This commit adds a command line switch `--videoMenu`. When set to true, a new menu entry called "Video" is added to the menu. This menu entry contains two sub entries that allow one to force PiP mode on shared screen videos or to toggle video controls on all video elements.