Make top bar elements clickable in frameless mode (#2009)
* Make top bar elements clickable in frameless mode

When frameless mode is activated the native title bar is no longer
rendered. It is up to the embedded website to define an area that
takes over its role using a dedicated CSS attribute. When this
attribute is set on the element, none of the click events are handled
in HTML code, but instead the drags and double clicks are handled in
the same way as for the native title bar. If the title bar element
contains items that must be clickable, they also need to have a
dedicated CSS attribute set.

In Teams the top bar with the search box is by default configured to
act as the title bar and the two buttons on the right - the settings
and account buttons - are by default clickable. This however means
that the remaining elements on the top bar are no longer clickable in
frameless mode, which prevents them from functioning as intended.

This change sets the correct "clickable" attribute on the waffle menu,
navigation buttons and the search box, so that they work correctly in
frameless mode.

Signed-off-by: Krzysztof Nowicki <krzysztof.a.nowicki@gmail.com>

* Apply suggestion from @IsmaelMartinez

---------

Signed-off-by: Krzysztof Nowicki <krzysztof.a.nowicki@gmail.com>
Co-authored-by: Krzysztof Nowicki <krzysztof.a.nowicki@gmail.com>
Co-authored-by: IsmaelMartinez <ISMAELMARTINEZ@GMAIL.COM>